+++ lib/model/game.dart
@@ -1,3 +1,6 @@
+import 'dart:typed_data';
+
+import 'package:dio/dio.dart';
 import 'package:game_lists/model/developer.dart';
 import 'package:game_lists/model/franchise.dart';
 import 'package:game_lists/model/genre.dart';
@@ -5,66 +8,90 @@ import 'package:game_lists/model/platform.dart';
 import 'package:game_lists/model/status.dart';
 import 'package:game_lists/model/walkthrough.dart';
 import 'package:hive/hive.dart';
+import 'package:image/image.dart';
 
 part 'game.g.dart';
+
+Future<Uint8List> getImageFromUrl(String url) async {
+  print(url);
+  Uint8List rawImage = Uint8List.fromList(
+      (await Dio().get(url, options: Options(responseType: ResponseType.bytes)))
+          .data);
+  var image = decodeImage(rawImage)!;
+  if (image.width > 240) {
+    var finalW = 240;
+    var finalH = ((finalW * image.height) / image.width).round();
+    image = copyResize(image,
+        width: finalW, height: finalH, interpolation: Interpolation.linear);
+  }
+  return Uint8List.fromList(encodePng(image));
+}
 
 @HiveType(typeId: 0)
 class Game extends HiveObject {
   @HiveField(0)
-  int giantBombId;
+  int? giantBombId;
 
   @HiveField(1)
-  DateTime dateLastUpdated;
+  int? steamId;
 
   @HiveField(2)
-  String name;
+  DateTime dateAdded;
 
   @HiveField(3)
-  String imageUrl;
+  DateTime dateLastUpdated;
 
   @HiveField(4)
-  String? description;
+  String name;
 
   @HiveField(5)
-  DateTime? releaseDate;
+  Uint8List? image;
 
   @HiveField(6)
-  List<Developer> developers;
+  String? description;
 
   @HiveField(7)
-  List<Franchise> franchises;
+  DateTime? releaseDate;
 
   @HiveField(8)
-  List<Genre> genres;
+  List<Developer> developers;
 
   @HiveField(9)
-  List<Platform> platforms;
+  List<Franchise> franchises;
 
   @HiveField(10)
-  int rating;
+  List<Genre> genres;
 
   @HiveField(11)
-  String notes;
+  List<Platform> platforms;
 
   @HiveField(12)
-  Status status;
+  int rating;
 
   @HiveField(13)
+  String notes;
+
+  @HiveField(14)
+  Status status;
+
+  @HiveField(15)
   List<Walkthrough> walkthroughs;
 
-  Game(
-      this.giantBombId,
-      this.dateLastUpdated,
-      this.name,
-      this.imageUrl,
-      this.description,
-      this.releaseDate,
-      this.developers,
-      this.franchises,
-      this.genres,
-      this.platforms,
-      this.rating,
-      this.notes,
-      this.status,
-      this.walkthroughs);
+  Game({
+    this.giantBombId,
+    this.steamId,
+    required this.dateLastUpdated,
+    required this.name,
+    this.image,
+    this.description,
+    this.releaseDate,
+    required this.developers,
+    required this.franchises,
+    required this.genres,
+    required this.platforms,
+    required this.rating,
+    required this.notes,
+    required this.status,
+    required this.walkthroughs,
+  }) : this.dateAdded = DateTime.now();
 }