+++ lib/model/genre.g.dart
@@ -17,7 +17,7 @@ class GenreAdapter extends TypeAdapter<Genre> {
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
     return Genre(
-      fields[0] as int,
+      fields[0] as int?,
       fields[1] as String,
     );
   }